+++ rasgotransforms/rasgotransforms/__init__.py
@@ -1,6 +1,6 @@
 from .main import serve_rasgo_transform_templates, DataWarehouse
 from .infer_columns import infer_columns
-from .jinja import render
+from .environment import Renderer
 from .transforms import Transforms
 
 transforms = Transforms()
